+++ lib/wired.dart
@@ -1,4 +1,4 @@
-library force_it;
+library wired;
 
 import 'dart:mirrors';
 import 'package:forcemirrors/force_mirrors.dart';
@@ -7,4 +7,3 @@ import 'package:yaml/yaml.dart';
 part 'src/annotations/annotations.dart';
 part 'src/di/bootstrap.dart';
 part 'src/di/messages_context.dart';
-part 'src/di/dependency_tree.dart';